+++ src/Main.cpp
@@ -153,10 +153,10 @@ void SS2K::maintenanceLoop(void *pvParameters) {
     }
     ss2k.lastShifterPosition = rtConfig.getShifterPosition();
 
-    if ((millis() - intervalTimer) > 20000) {  // add check here for when to restart WiFi
+    if ((millis() - intervalTimer) > 10000) {  // add check here for when to restart WiFi
                                                // maybe if in STA mode and 8.8.8.8 no ping return?
       // ss2k.restartWifi();
-
+      ss2k.toggleUDPLog(false);  // check to see if UDP Logging is disabled
       intervalTimer = millis();
     }
 
@@ -176,6 +176,25 @@ void SS2K::maintenanceLoop(void *pvParameters) {
   }
 }
 #endif  // UNIT_TEST
+
+// Test is False, Request Off is True
+void SS2K::toggleUDPLog(bool requestOff) {
+  static bool userUDPpreference = userConfig.getUdpLogEnabled();
+  if (!ss2k.UDPloggingAutoDisabled && !requestOff) {  // timer check only, return fast.
+    return;
+  }
+  if (!ss2k.UDPloggingAutoDisabled && requestOff) {  // Initial request off
+    userConfig.setUdpLogEnabled(false);
+    ss2k.UDPloggingAutoDisabled = true;
+  }
+  if (ss2k.UDPloggingAutoDisabled && requestOff) {  // Repeat request off
+    userConfig.setUdpLogEnabled(false);
+  }
+  if (ss2k.UDPloggingAutoDisabled && !requestOff) {  // timer expired, reset to original user setting.
+    userConfig.setUdpLogEnabled(userUDPpreference);
+    ss2k.UDPloggingAutoDisabled = false;
+  }
+}
 
 void SS2K::restartWifi() {
   httpServer.stop();
@@ -261,7 +280,7 @@ bool IRAM_ATTR SS2K::deBounce() {
 
 ///////////// Interrupt Functions /////////////
 void IRAM_ATTR SS2K::shiftUp() {  // Handle the shift up interrupt IRAM_ATTR is to keep the interrput code in ram always
-  if (ss2k.deBounce()) {
+  if (ss2k.deBounce() && !rtConfig.getERGMode()) {
     if (!digitalRead(SHIFT_UP_PIN)) {  // double checking to make sure the interrupt wasn't triggered by emf
       rtConfig.setShifterPosition(rtConfig.getShifterPosition() - 1 + userConfig.getShifterDir() * 2);
     } else {
@@ -271,7 +290,7 @@ void IRAM_ATTR SS2K::shiftUp() {  // Handle the shift up interrupt IRAM_ATTR is 
 }
 
 void IRAM_ATTR SS2K::shiftDown() {  // Handle the shift down interrupt
-  if (ss2k.deBounce()) {
+  if (ss2k.deBounce() && !rtConfig.getERGMode()) {
     if (!digitalRead(SHIFT_DOWN_PIN)) {  // double checking to make sure the interrupt wasn't triggered by emf
       rtConfig.setShifterPosition(rtConfig.getShifterPosition() + 1 - userConfig.getShifterDir() * 2);
     } else {